+++ app/src/main/java/com/example/itplaneta/ui/navigation/Screens.kt
@@ -4,7 +4,7 @@ sealed class Screens(val route  : String){
     object Main : Screens("main_screen")
     object AddAccount : Screens("account_screen")
     object EditAccount : Screens("account_screen/{accountId}"){
-        fun passAccountId(accountId: Long) = "account_screen/$accountId"
+        fun passAccountId(accountId: Int) = "account_screen/$accountId"
     }
     object QrScanner : Screens("qrscanner_screen")
 }
